for Docker

+++ src/main/java/ru/kpfu/itis/kononenko/entity/MergeRequest.java
@@ -3,10 +3,10 @@ package ru.kpfu.itis.kononenko.entity;
 import java.sql.Timestamp;
 
 public record MergeRequest(
-        int id,
-        int requesterTreeId,
-        int targetTreeId,
-        int commonAncestorId,
+        Long id,
+        Long requesterTreeId,
+        Long targetTreeId,
+        Long commonAncestorId,
         String status,
         Timestamp createdAt
 ) {}
